+++ src/components/contact/contact.config.js
@@ -1,0 +1,7 @@
+export default {
+  element: 'section',
+  attributes: {
+    class: 'contact',
+  },
+  children: [{}],
+};