+++ src/components/loading/loading.js
@@ -1,0 +1,8 @@
+export default () => {
+  const loading = {
+    init() {},
+    cacheDOM() {},
+    bindEvents() {},
+    render() {},
+  };
+};