+++ src/components/navbar/navbar.config.js
@@ -1,0 +1,49 @@
+export default {
+  element: 'nav',
+  children: [
+    {
+      element: 'ul',
+      attributes: {
+        class: 'nav_left',
+      },
+      children: [
+        {
+          element: 'li',
+          children: [
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    src: '#',
+                  },
+                },
+                {
+                  element: 'h1',
+                  attributes: {
+                    textContent: 'John Doe',
+                  },
+                },
+              ],
+            },
+          ],
+        },
+      ],
+    },
+    {
+      element: 'ul',
+      attributes: {
+        class: 'nav_right',
+      },
+      children: [
+        {
+          element: 'li',
+        },
+      ],
+    },
+  ],
+};