+++ src/components/projects/projects.config.js
@@ -1,0 +1,11 @@
+export default {
+  element: 'section',
+  attributes: {
+    class: 'projects',
+  },
+  children: [
+    {
+      element: 'article',
+    },
+  ],
+};