+++ src/helpers/createElement.js
@@ -1,0 +1,48 @@
+const BuildElement = (state) => ({
+  setAttributes: (attributes) => {
+    Object.entries(attributes).forEach(([key, value]) => {
+      if (key !== 'textContent') {
+        if (key === 'class') {
+          state.setClassName(value.split(/\s/));
+        } else if (key === 'style') {
+          state.setStyle(value);
+        } else {
+          state.setAttribute(key, value);
+        }
+      } else {
+        state.setTextContent(value);
+      }
+    });
+  },
+  setStyle: (text) => {
+    state.style.cssText = text;
+  },
+  setID: (id) => {
+    state.id = id;
+  },
+  setClassName: (arrClass) => {
+    arrClass.forEach((className) => state.classList.add(className));
+  },
+  setTextContent: (text) => {
+    state.textContent = text;
+  },
+  setChildren: (children) => {
+    children.forEach((child) => {
+      const childElement = createElement(child.element);
+      if (child.attributes && child.attributes.constructor.name === 'Object') {
+        childElement.setAttributes(child.attributes);
+      }
+      if (child.children) {
+        childElement.setChildren(child.children);
+      }
+      state.appendChild(childElement);
+    });
+  },
+});
+
+export default function createElement(tagName, attributes) {
+  const htmlElement = document.createElement(tagName);
+  Object.assign(htmlElement, BuildElement(htmlElement));
+  if (attributes) htmlElement.setAttributes(attributes);
+  return htmlElement;
+}