+++ src/helpers/pubSub.js
@@ -1,0 +1,20 @@
+export default {
+  subscribers: {},
+  subscribe(subscriber, fn) {
+    this.subscribers[subscriber] = this.subscribers[subscriber] || [];
+    if (!this.subscribers[subscriber].some((handler) => handler.name === fn.name)) {
+      this.subscribers[subscriber].push(fn);
+    }
+  },
+  unsubsccribe(subscriber, fn) {
+    if (this.subscribers[subscriber]) {
+      this.subscribers[subscriber].splice(this.subscribers[subscriber].indexOf(fn), 1);
+      if (this.subscribers[subscriber].length === 0) delete this.subscribers[subscriber];
+    }
+  },
+  publish(subscriber, ...args) {
+    if (this.subscribers[subscriber]) {
+      this.subscribers[subscriber].forEach((fn) => fn(...args));
+    }
+  },
+};