+++ webpack.config.js
@@ -4,15 +4,12 @@ const path = require('path');
 module.exports = {
   mode: 'development',
   entry: {
-    index: './src/index.js',
+    index: './src/app.js',
   },
   devtool: 'inline-source-map',
-  devServer: {
-    static: './dist',
-  },
   plugins: [
     new HtmlWebpackPlugin({
-      title: 'Module Webpack Starter',
+      title: 'Homepage',
     }),
   ],
   output: {
@@ -40,4 +37,4 @@ module.exports = {
   optimization: {
     runtimeChunk: 'single',
   },
-}
+};